get pages working again, still needs refactoring
