+++ webapp/app/components/ViewMasterReel.tsx
@@ -1,16 +1,21 @@
-import type { Project } from "../data/getProjects";
+import type { Project } from "../data/getProjects.server";
 import { useMarkdown } from "../hooks/useMarkdown";
 import { formatDate } from "../util/formatDate";
 import { formatArtMediumIdToText } from "../util/formatArtMediumIdToText";
 import { ProjectLink } from "./ProjectLink";
+import { getInstagramUrl } from "../util/getInstagramUrl";
+import { getRecordId } from "../util/getRecordId";
 
 type ViewMasterReelProps = {
   reel: Project;
 };
 export function ViewMasterReel({ reel }: ViewMasterReelProps) {
-  const { title, type, author, date, body, id, externalHref, images } = reel;
+  const { title, type, author, date, body, externalHref, instagramId, images } =
+    reel;
+  const { id } = getRecordId(reel.id);
   const bodyHtml = useMarkdown(body);
   const gridRows = (images?.length || 0) > 2 ? "grid-rows-2" : "grid-rows-1";
+  const href = instagramId ? getInstagramUrl(instagramId) : externalHref;
   return (
     <div className="pb-4 view-master-reel">
       <div className="flex flex-col sm:flex-row">
@@ -43,7 +48,7 @@ export function ViewMasterReel({ reel }: ViewMasterReelProps) {
             by: {author}, {formatDate(new Date(date))}
           </div>
           {bodyHtml}
-          <ProjectLink externalHref={externalHref}>
+          <ProjectLink externalHref={href}>
             {formatArtMediumIdToText(id, type)}
           </ProjectLink>
         </div>
