+++ webapp/app/routes/seeds.no-1.tsx
@@ -54,7 +54,7 @@ export default function SeedNo1() {
           <div className="mt-16">
             <Print
               print={{
-                id: "presentation-no-1",
+                id: "uncooked:⟨presentation-no-1⟩",
                 type: "presentation",
                 title: "Understanding",
                 author: "Austin Trautman",
