+++ webapp/app/util/getInstagramUrl.ts
@@ -1,0 +1,3 @@
+export function getInstagramUrl(id: string): string {
+  return `https://www.instagram.com/p/${id}`;
+}
